try fix windows

+++ demo/c/test/test_orca_c.py
@@ -79,7 +79,7 @@ class OrcaCTestCase(unittest.TestCase):
         os.remove(output_path)
 
     def run_orca_streaming(self, model_path: str) -> None:
-        output_path = os.path.join(os.path.dirname(__file__), "output.wav")
+        output_path = os.path.join(os.path.dirname(__file__), "output_streaming.wav")
         args = [
             os.path.join(os.path.dirname(__file__), "../build/orca_demo_streaming"),
             "-a", self._access_key,
@@ -92,7 +92,13 @@ class OrcaCTestCase(unittest.TestCase):
         process = subprocess.Popen(args, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
         stdout, stderr = process.communicate()
 
-        self.assertEqual(process.poll(), 0)
+        poll_result = process.poll()
+        if poll_result != 0:
+            print(stdout.decode('utf-8'))
+            print(stderr.decode('utf-8'))
+            raise RuntimeError("Error running demo. See details above")
+
+        self.assertEqual(poll_result, 0)
         self.assertEqual(stderr.decode('utf-8'), '')
         self.assertTrue("Saved final audio" in stdout.decode('utf-8'))
         os.remove(output_path)
